+++ software/firmware/lib/include/stdint_fast.h
